requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Image Processing
Pillow>=10.2.0
opencv-python>=4.9.0
numpy>=1.26.0

# AI/ML (PyTorch with CUDA)
# Note: Install PyTorch separately with CUDA support
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu126
diffusers>=0.30.1
transformers>=4.46.2
safetensors>=0.4.0
omegaconf>=2.3.0
accelerate>=0.25.0
einops>=0.7.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
httpx>=0.26.0

# WebSocket
websockets>=12.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
huggingface-hub>=0.20.0

# ONNX Runtime (for Pose detection)
onnxruntime>=1.17.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
